+++ tensorflow/merlintf/keras/layers/convolutional/conv2dt.py
@@ -415,7 +415,7 @@ class Conv2dtTest(unittest.TestCase):
             model = Conv2Dt(nf_out, kernel_size=ksz, shapes=shape, axis_conv_t=2, intermediate_filters=nf_inter,
                             strides=stride, data_format=data_format, use_3D_convs=use_3D_convs)
 
-        x = tf.cast(tf.random.normal(shape), dtype=tf.float32)
+        x = tf.random.normal(shape)
         Kx = model(x)
 
         self.assertTrue(Kx.shape == expected_shape)
