+++ tensorflow/merlintf/keras/models/unet.py
@@ -638,19 +638,19 @@ class UNetTest(unittest.TestCase):
 
         if dim == '2D':
             if complex_input:
-                x = merlintf.random_normal_complex((nBatch, M, N, 1), dtype=tf.float32)
+                x = merlintf.random_normal_complex((nBatch, M, N, 1))
             else:
-                x = tf.random.normal((nBatch, M, N, 1), dtype=tf.float32)
+                x = tf.random.normal((nBatch, M, N, 1))
         elif dim == '3D' or dim == '2Dt':
             if complex_input:
-                x = merlintf.random_normal_complex((nBatch, M, N, D, 1), dtype=tf.float32)
+                x = merlintf.random_normal_complex((nBatch, M, N, D, 1))
             else:
-                x = tf.random.normal((nBatch, M, N, D, 1), dtype=tf.float32)
+                x = tf.random.normal((nBatch, M, N, D, 1))
         elif dim == '3Dt':
             if complex_input:
-                x = merlintf.random_normal_complex((nBatch, T, M, N, D, 1), dtype=tf.float32)
+                x = merlintf.random_normal_complex((nBatch, T, M, N, D, 1))
             else:
-                x = tf.random.normal((nBatch, T, M, N, D, 1), dtype=tf.float32)
+                x = tf.random.normal((nBatch, T, M, N, D, 1))
         else:
             raise RuntimeError(f'No implementation for dim {dim} available!')
